Clarify which checks are enabled.
Since one of the goals of this analyzer is to verify that a lookout setup is
working, it helps to know which checks are enabled. Otherwise, it's hard to
distinguish a broken setup from a PR that simply has no affecting changes.
To make it easier to tell, enumerate the enabled checks explicitly.

Addresses part of https://github.com/src-d/empathy-sessions/issues/31

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>